Add float to string function

+++ native_template/source/softfloat.c
@@ -36,6 +36,7 @@
 */
 
 #include "softfloat.h"
+#include "f2d.h"
 
 /*----------------------------------------------------------------------------
  | Returns 1 if the single-precision floating-point value `a' is a signaling
@@ -565,7 +566,7 @@ unsigned int getfraction(unsigned int shifts, unsigned int fleadzeros, unsigned 
 	return out;
 }
 
-unsigned int ASCII_to_float32(char* in) //Copyright 2014 Eugene Rockey
+float32 ASCII_to_float32(char* in) //Copyright 2014 Eugene Rockey
 {									//converts ASCII decimal numbers to IEEE-754 32-bit unsigned int
 	unsigned int out = 0, index = 0, integer = 0, fraction = 0, biasedexp = 0, shifts = 0, fleadzeros = 0, ibits = 0;
